+++ packages/artifact/__tests__/download.test.ts
@@ -12,8 +12,12 @@ import {
   ListArtifactsResponse,
   QueryArtifactResponse
 } from '../src/internal/contracts'
+import * as stream from 'stream'
+import {gzip} from 'zlib'
+import {promisify} from 'util'
 
 const root = path.join(__dirname, '_temp', 'artifact-download-tests')
+const defaultEncoding = 'utf8'
 
 jest.mock('../src/internal/config-variables')
 jest.mock('@actions/http-client')
@@ -67,7 +71,7 @@ describe('Download Tests', () => {
     setupFailedResponse()
     const downloadHttpClient = new DownloadHttpClient()
     expect(downloadHttpClient.listArtifacts()).rejects.toThrow(
-      'Unable to list artifacts for the run'
+      'List Artifacts failed: Artifact service responded with 400'
     )
   })
 
@@ -109,58 +113,125 @@ describe('Download Tests', () => {
         configVariables.getRuntimeUrl()
       )
     ).rejects.toThrow(
-      `Unable to get ContainersItems from ${configVariables.getRuntimeUrl()}`
+      `Get Container Items failed: Artifact service responded with 400`
     )
   })
 
   it('Test downloading an individual artifact with gzip', async () => {
-    setupDownloadItemResponse(true, 200)
+    const fileContents = Buffer.from(
+      'gzip worked on the first try\n',
+      defaultEncoding
+    )
+    const targetPath = path.join(root, 'FileA.txt')
+
+    setupDownloadItemResponse(fileContents, true, 200, false, false)
     const downloadHttpClient = new DownloadHttpClient()
 
     const items: DownloadItem[] = []
     items.push({
       sourceLocation: `${configVariables.getRuntimeUrl()}_apis/resources/Containers/13?itemPath=my-artifact%2FFileA.txt`,
-      targetPath: path.join(root, 'FileA.txt')
+      targetPath
     })
 
     await expect(
       downloadHttpClient.downloadSingleArtifact(items)
     ).resolves.not.toThrow()
+
+    await checkDestinationFile(targetPath, fileContents)
   })
 
   it('Test downloading an individual artifact without gzip', async () => {
-    setupDownloadItemResponse(false, 200)
+    const fileContents = Buffer.from(
+      'plaintext worked on the first try\n',
+      defaultEncoding
+    )
+    const targetPath = path.join(root, 'FileB.txt')
+
+    setupDownloadItemResponse(fileContents, false, 200, false, false)
     const downloadHttpClient = new DownloadHttpClient()
 
     const items: DownloadItem[] = []
     items.push({
       sourceLocation: `${configVariables.getRuntimeUrl()}_apis/resources/Containers/13?itemPath=my-artifact%2FFileB.txt`,
-      targetPath: path.join(root, 'FileB.txt')
+      targetPath
     })
 
     await expect(
       downloadHttpClient.downloadSingleArtifact(items)
     ).resolves.not.toThrow()
+
+    await checkDestinationFile(targetPath, fileContents)
   })
 
   it('Test retryable status codes during artifact download', async () => {
     // The first http response should return a retryable status call while the subsequent call should return a 200 so
     // the download should successfully finish
-    const retryableStatusCodes = [429, 502, 503, 504]
+    const retryableStatusCodes = [429, 500, 502, 503, 504]
     for (const statusCode of retryableStatusCodes) {
-      setupDownloadItemResponse(false, statusCode)
+      const fileContents = Buffer.from('try, try again\n', defaultEncoding)
+      const targetPath = path.join(root, `FileC-${statusCode}.txt`)
+
+      setupDownloadItemResponse(fileContents, false, statusCode, false, true)
       const downloadHttpClient = new DownloadHttpClient()
 
       const items: DownloadItem[] = []
       items.push({
         sourceLocation: `${configVariables.getRuntimeUrl()}_apis/resources/Containers/13?itemPath=my-artifact%2FFileC.txt`,
-        targetPath: path.join(root, 'FileC.txt')
+        targetPath
       })
 
       await expect(
         downloadHttpClient.downloadSingleArtifact(items)
       ).resolves.not.toThrow()
+
+      await checkDestinationFile(targetPath, fileContents)
     }
+  })
+
+  it('Test retry on truncated response with gzip', async () => {
+    const fileContents = Buffer.from(
+      'Sometimes gunzip fails on the first try\n',
+      defaultEncoding
+    )
+    const targetPath = path.join(root, 'FileD.txt')
+
+    setupDownloadItemResponse(fileContents, true, 200, true, true)
+    const downloadHttpClient = new DownloadHttpClient()
+
+    const items: DownloadItem[] = []
+    items.push({
+      sourceLocation: `${configVariables.getRuntimeUrl()}_apis/resources/Containers/13?itemPath=my-artifact%2FFileD.txt`,
+      targetPath
+    })
+
+    await expect(
+      downloadHttpClient.downloadSingleArtifact(items)
+    ).resolves.not.toThrow()
+
+    await checkDestinationFile(targetPath, fileContents)
+  })
+
+  it('Test retry on truncated response without gzip', async () => {
+    const fileContents = Buffer.from(
+      'You have to inspect the content-length header to know if you got everything\n',
+      defaultEncoding
+    )
+    const targetPath = path.join(root, 'FileE.txt')
+
+    setupDownloadItemResponse(fileContents, false, 200, true, true)
+    const downloadHttpClient = new DownloadHttpClient()
+
+    const items: DownloadItem[] = []
+    items.push({
+      sourceLocation: `${configVariables.getRuntimeUrl()}_apis/resources/Containers/13?itemPath=my-artifact%2FFileD.txt`,
+      targetPath
+    })
+
+    await expect(
+      downloadHttpClient.downloadSingleArtifact(items)
+    ).resolves.not.toThrow()
+
+    await checkDestinationFile(targetPath, fileContents)
   })
 
   /**
@@ -226,52 +297,111 @@ describe('Download Tests', () => {
    * @param firstHttpResponseCode the http response code that should be returned
    */
   function setupDownloadItemResponse(
+    fileContents: Buffer,
     isGzip: boolean,
-    firstHttpResponseCode: number
+    firstHttpResponseCode: number,
+    truncateFirstResponse: boolean,
+    retryExpected: boolean
   ): void {
-    jest
-      .spyOn(DownloadHttpClient.prototype, 'pipeResponseToFile')
-      .mockImplementationOnce(async () => {
-        return new Promise<void>(resolve => {
-          resolve()
-        })
-      })
-
-    jest
+    const spyInstance = jest
       .spyOn(HttpClient.prototype, 'get')
       .mockImplementationOnce(async () => {
-        const mockMessage = new http.IncomingMessage(new net.Socket())
-        mockMessage.statusCode = firstHttpResponseCode
-        if (isGzip) {
-          mockMessage.headers = {
-            'content-type': 'gzip'
+        if (firstHttpResponseCode === 200) {
+          const fullResponse = await constructResponse(isGzip, fileContents)
+          const actualResponse = truncateFirstResponse
+            ? fullResponse.subarray(0, 3)
+            : fullResponse
+
+          return {
+            message: getDownloadResponseMessage(
+              firstHttpResponseCode,
+              isGzip,
+              fullResponse.length,
+              actualResponse
+            ),
+            readBody: emptyMockReadBody
+          }
+        } else {
+          return {
+            message: getDownloadResponseMessage(
+              firstHttpResponseCode,
+              false,
+              0,
+              null
+            ),
+            readBody: emptyMockReadBody
           }
         }
-
-        return new Promise<HttpClientResponse>(resolve => {
-          resolve({
-            message: mockMessage,
-            readBody: emptyMockReadBody
-          })
-        })
       })
-      .mockImplementationOnce(async () => {
+
+    // set up a second mock only if we expect a retry. Otherwise this mock will affect other tests.
+    if (retryExpected) {
+      spyInstance.mockImplementationOnce(async () => {
         // chained response, if the HTTP GET function gets called again, return a successful response
-        const mockMessage = new http.IncomingMessage(new net.Socket())
-        mockMessage.statusCode = 200
-        if (isGzip) {
-          mockMessage.headers = {
-            'content-type': 'gzip'
-          }
+        const fullResponse = await constructResponse(isGzip, fileContents)
+        return {
+          message: getDownloadResponseMessage(
+            200,
+            isGzip,
+            fullResponse.length,
+            fullResponse
+          ),
+          readBody: emptyMockReadBody
         }
-
-        return new Promise<HttpClientResponse>(resolve => {
-          resolve({
-            message: mockMessage,
-            readBody: emptyMockReadBody
-          })
-        })
       })
+    }
+  }
+
+  async function constructResponse(
+    isGzip: boolean,
+    plaintext: Buffer | string
+  ): Promise<Buffer> {
+    if (isGzip) {
+      return await promisify(gzip)(plaintext)
+    } else if (typeof plaintext === 'string') {
+      return Buffer.from(plaintext, defaultEncoding)
+    } else {
+      return plaintext
+    }
+  }
+
+  function getDownloadResponseMessage(
+    httpResponseCode: number,
+    isGzip: boolean,
+    contentLength: number,
+    response: Buffer | null
+  ): http.IncomingMessage {
+    let readCallCount = 0
+    const mockMessage = <http.IncomingMessage>new stream.Readable({
+      read(size) {
+        switch (readCallCount++) {
+          case 0:
+            if (!!response && response.byteLength > size) {
+              throw new Error(
+                `test response larger than requested size (${size})`
+              )
+            }
+            this.push(response)
+            break
+
+          default:
+            // end the stream
+            this.push(null)
+            break
+        }
+      }
+    })
+
+    mockMessage.statusCode = httpResponseCode
+    mockMessage.headers = {
+      'content-length': contentLength.toString()
+    }
+
+    if (isGzip) {
+      mockMessage.headers['content-encoding'] = 'gzip'
+    }
+
+    return mockMessage
   }
 
   /**
@@ -338,7 +468,7 @@ describe('Download Tests', () => {
   function setupFailedResponse(): void {
     jest.spyOn(HttpClient.prototype, 'get').mockImplementationOnce(async () => {
       const mockMessage = new http.IncomingMessage(new net.Socket())
-      mockMessage.statusCode = 500
+      mockMessage.statusCode = 400
       return new Promise<HttpClientResponse>(resolve => {
         resolve({
           message: mockMessage,
@@ -346,5 +476,15 @@ describe('Download Tests', () => {
         })
       })
     })
+  }
+
+  async function checkDestinationFile(
+    targetPath: string,
+    expectedContents: Buffer
+  ): Promise<void> {
+    const fileContents = await fs.readFile(targetPath)
+
+    expect(fileContents.byteLength).toEqual(expectedContents.byteLength)
+    expect(fileContents.equals(expectedContents)).toBeTruthy()
   }
 })