+++ packages/cache/src/internal/cacheUtils.ts
@@ -35,7 +35,7 @@ export async function createTempDirectory(): Promise<string> {
   return dest
 }
 
-export function getArchiveFileSizeIsBytes(filePath: string): number {
+export function getArchiveFileSizeInBytes(filePath: string): number {
   return fs.statSync(filePath).size
 }
 
@@ -47,7 +47,9 @@ export async function resolvePaths(patterns: string[]): Promise<string[]> {
   })
 
   for await (const file of globber.globGenerator()) {
-    const relativeFile = path.relative(workspace, file)
+    const relativeFile = path
+      .relative(workspace, file)
+      .replace(new RegExp(`\\${path.sep}`, 'g'), '/')
     core.debug(`Matched: ${relativeFile}`)
     // Paths are made relative so the tar entries are all relative to the root of the workspace.
     paths.push(`${relativeFile}`)